Create Readme.md
This project is a simple Handwriting to Text Conversion. You just need to give the images of handwritten text as input, and the extracted text in digital form will be the output.
This will help those people who want to extract the text from images to complete their notes rapidly. This project is built using Python3.